==> 10 - ML Trial 4 new.ipynb <==
# BA prediction ML Trial 4

## Summary up to this point:
### Previous:
* __"1D Feature Experiments:"__
  * The _"1D Features"_ refers to the original experiments based on quantitative (and typically, whole-molecule) calculations in RDKit for each drug... e.g., MolWt, LogP, NumHDonors, NumHAcceptors, etc. 
  * Of the many RDKit calculations available as features, we tested performance of linear regression model according to various feature sets -- the feature sets represented different subsets of the RDKit calculations
    * feature set 1 (1a,1b) had all RdKit calculations, only differed in whether VSA was represented as absolute value of VSA _versus_ percent-VSA of total VSA.
    * feature set 4 had the fewest number of calculations, focusing only on those reported in literature as being strongly tied to bioavailability; these are the metrics commonly used as filters in screening
    * feature set 2 was a copy of feature set 1, but narrowed by running set 1 thru a cross-correlation matrix and removing variables with correlation > 0.95
    * there were also ANOVA variants of the feature sets, which used an ANOVA test against a 3-category label (low BA, mid BA, high BA), and deleted features which did not meet statistical significance criteria (i.e., features for which no statistically significant difference existed in that variable for at least 1 of the categories)
  * Regarding feature selection, __Feature set 1b__ was the best performing RDKit feature subset, followed most closely by 1b-ANOVA.
  * Regarding labeling scheme, only _Continuous_ (regression) and _3-category_ (classification) were tested.
    * for regression, the best r2 value (i.e. feature set 1b) was typically between 0.14-0.20 in Linear Regression models.
    * for classification (3-category), accuracy was somewhere around 0.5 
<br><br>
* __"Fragment experiments:"__
  * Hypothesis: 
    1. it is hypothesized that the model does not predict well __because it covers many drug types / classes, which are not significantly differentiated by the features__. 
    2. it is further hypothesized that the missing information about drug classes _could be_ tied to its 2D structure - or more specifically, tied to the presence/absence of key structural motifs / scaffolds in the chemical structure of various drug classes. _(For a simple example of a link, a drug with acidic/basic groups allow its preparation as a pharmaceutical salt, boosting bioavailability.)_
    <br>
  * Thus, it was decided to attempt a representation of drug classes and chemical structure using information about their 2D chemical structure - _specifically_, using an array of chemical fragments. 
    * Each drug was subjected to fragmentation, creating a fragment library.
    * Each drug's Murcko Scaffold, and fragmentation products thereof, were added to the fragment library.
    * Drugs were linked to the fragments, scaffolds, and scaffold fragments by explicit lookup between drug & fragment library, _but were also_ linked indirectly by searching for Substructure Matches across all drugs/fragments. <br><br>
  * __Latest finding:__ 
    * The fragment set having most correlation with bioavailability was __`frags_all`__, which is the comprehensive set based on both fragment production _and_ matches of substructure search.  
    * The NLP technique best representing the generated fragments was: (1) using __CountVectorizer__ on fragment array... (2) __TF-IDF__... (3) using __n-gram (n=2) + CountVectorizer__.... followed by (4) word2vec

### Agenda for this NB:

* Determine the best Regression Model for BA prediction based on fragments
* Determine the best Classification Model for BA prediction based on fragments
* Eventually:
  * compare performance of fragments when used in conjunction with original 1D feature sets

## UPDATED Agenda for this NB:

* Initially, an attempt was made to combine the fragment NLP features w/ the RDKit "1D" features. The resulting r2 value was terrible (negative); this is because the numerical RDKIT features become rather useless once they've been joined into the large / high-dimensionality vector high of the NLP features.
* __To resolve this problem,__ 
  * We will use NLP features and the Numerical RDKit calculations separately, instead of combining.
  * Essentially, the new idea will be to: 
    - Run a Clasification Model using the NLP features to classify the BA of each drug (use cat labels)
    - Run a Regression Model using numerical features _and predicted class_ to predict final bioavailability 

### <font color='red'> Actual Steps: </font>
* <font color='green'>__Step 1.__</font> verify this approach is feasible using RFC w/ NLP features (__"Model 1"__) followed by RFR w/ Numerical features (__"Model 2"__) <br><br>
* <font color='green'>__Step 2(a).__</font> if feasible (as per above), then determine the best model type for _Model 1_ (classification step) 
* <font color='green'>__Step 2(b).__</font> does the choice of categorical label affect which model appears best model for classification _Model 1_ ?  <br> _(e.g., label_cat0, label_cat1, label_cat2, label_cat3)_ <br><br>

* <font color='green'>__Step 3.__</font> Run trials comparing the combination of Model 1 and Model 2. <br> __Variables to study include:__
  * Classification as 1, Regression as 2 <br> _vs._ Regression as 1, Classification as 2 <br> _vs._ Regression as 1 and 2 <br> _vs._ Classification as 1 and 2 
  * number of classes for Classification _(e.g., label_cat0, label_cat1, label_cat2, label_cat3)_
  * Labels for Regression Model 2: predict absolute BA value..? or predict ___residual BA___ from model 1?
  * Manual combination of Model 1/Model 2 predictions, <br> _versus_ direct use of Model 1 prediction in Model 2 features?

# MSM VM config prep
import findspark
findspark.init('/home/mitch/spark-3.3.0-bin-hadoop2')
import pyspark
 
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName('BApredsV4').getOrCreate()

# --- suppress future spark warnings/error/etc output ---
spark.sparkContext.setLogLevel("OFF")

import pandas as pd
data = pd.read_pickle("data_final_NEW_2022-09-05_aspandas.pkl")
pd.set_option('display.max_columns',None)
data = data.rename(columns={'Name':'name','_c0':'index'})
data = data.drop(columns='drug_name')
data = spark.createDataFrame(data)

import pandas as pd
#data = pd.read_pickle("bioavailability_data_masterFinal.pkl")
data = pd.read_pickle("bioavailability_data_masterFinal2.pkl")
pd.set_option('display.max_columns',None)
data = data.rename(columns={'Name':'name','_c0':'index'})
data = data.drop(columns='drug_name')
data = spark.createDataFrame(data)

data.printSchema()

## List of column titles

data.printSchema()

final_column_order = ['index','name','drug_smiles',                    
    # labels
    'label_q0','label_cat0','label_cat1','label_cat2','label_cat3','label_cat4',
                      
    # fragment features
    'FEAT_all_n2g_cv2','FEAT_subset2_cv5','FEAT_subset_n2g_cv2',
                      
    # rdkit features
    'FEAT_rdkit_1b','FEAT_rdkit_2b','FEAT_rdkit_1bANOVA','FEAT_rdkit_2bANOVA',
                      
    # fragment data
    'frags_all','frags_subset','frags_subset2','all_n2g_cv2','subset2_cv5','subset_n2g_cv2',
                      
    # rdkit data
    'MolWt','ExactMolWt','qed','MolLogP','MolMR','VSA_total','LabuteASA',
    'TPSA','MaxPartialCharge','MinPartialCharge','MaxAbsPartialCharge','MinAbsPartialCharge','NumHAcceptors',
    'NumHDonors','HeavyAtomCount','NumHeteroatoms','NumRotatableBonds','NHOHCount','NOCount','FractionCSP3',
    'RingCount','NumAliphaticRings','NumAromaticRings','NumAliphaticHeterocycles','NumAromaticHeterocycles',
    'NumSaturatedHeterocycles','NumSaturatedRings','BalabanJ','BertzCT','HallKierAlpha','fracVSA_PEOE01',
    'fracVSA_PEOE02','fracVSA_PEOE03','fracVSA_PEOE04','fracVSA_PEOE05','fracVSA_PEOE06','fracVSA_PEOE07',
    'fracVSA_PEOE08','fracVSA_PEOE09','fracVSA_PEOE10','fracVSA_PEOE11','fracVSA_PEOE12','fracVSA_PEOE13',
    'fracVSA_PEOE14','fracVSA_SMR01','fracVSA_SMR02','fracVSA_SMR03','fracVSA_SMR04','fracVSA_SMR05',
    'fracVSA_SMR06','fracVSA_SMR07','fracVSA_SMR08','fracVSA_SMR09','fracVSA_SMR10','fracVSA_SlogP01',
    'fracVSA_SlogP02','fracVSA_SlogP03','fracVSA_SlogP04','fracVSA_SlogP05','fracVSA_SlogP06','fracVSA_SlogP07',
    'fracVSA_SlogP08','fracVSA_SlogP09','fracVSA_SlogP10','fracVSA_SlogP11','fracVSA_SlogP12']

data_masterFinal_clean = data.select(final_column_order)

final_column_order = ['index','name','drug_smiles',                    
    # labels
    'label_q0','label_cat0','label_cat1','label_cat2','label_cat3','label_cat4',
                      
    # fragment features
    'FEAT_f_all_cv2', 'FEAT_f_all_cv2_idf', 'FEAT_f_all_cv5','FEAT_f_all_n2g_cv2', 'FEAT_f_all_n2g_cv5', 
    'FEAT_f_subset_cv2', 'FEAT_f_subset_cv2_idf', 'FEAT_f_subset_cv5', 'FEAT_f_subset_n2g_cv2', 'FEAT_f_subset_n2g_cv5', 
    'FEAT_f_subset2_cv2', 'FEAT_f_subset2_cv2_idf', 'FEAT_f_subset2_cv5', 'FEAT_f_subset2_n2g_cv2', 'FEAT_f_subset2_n2g_cv5',
                      
    # rdkit features
    'FEAT_rdkit_1b','FEAT_rdkit_2b','FEAT_rdkit_1bANOVA','FEAT_rdkit_2bANOVA',
                      
    # fragment data
    #'frags_all','frags_subset','frags_subset2','all_n2g_cv2','subset2_cv5','subset_n2g_cv2',
                      
    # rdkit data
    'MolWt','ExactMolWt','qed','MolLogP','MolMR','VSA_total','LabuteASA',
    'TPSA','MaxPartialCharge','MinPartialCharge','MaxAbsPartialCharge','MinAbsPartialCharge','NumHAcceptors',
    'NumHDonors','HeavyAtomCount','NumHeteroatoms','NumRotatableBonds','NHOHCount','NOCount','FractionCSP3',
    'RingCount','NumAliphaticRings','NumAromaticRings','NumAliphaticHeterocycles','NumAromaticHeterocycles',
    'NumSaturatedHeterocycles','NumSaturatedRings','BalabanJ','BertzCT','HallKierAlpha','fracVSA_PEOE01',
    'fracVSA_PEOE02','fracVSA_PEOE03','fracVSA_PEOE04','fracVSA_PEOE05','fracVSA_PEOE06','fracVSA_PEOE07',
    'fracVSA_PEOE08','fracVSA_PEOE09','fracVSA_PEOE10','fracVSA_PEOE11','fracVSA_PEOE12','fracVSA_PEOE13',
    'fracVSA_PEOE14','fracVSA_SMR01','fracVSA_SMR02','fracVSA_SMR03','fracVSA_SMR04','fracVSA_SMR05',
    'fracVSA_SMR06','fracVSA_SMR07','fracVSA_SMR08','fracVSA_SMR09','fracVSA_SMR10','fracVSA_SlogP01',
    'fracVSA_SlogP02','fracVSA_SlogP03','fracVSA_SlogP04','fracVSA_SlogP05','fracVSA_SlogP06','fracVSA_SlogP07',
    'fracVSA_SlogP08','fracVSA_SlogP09','fracVSA_SlogP10','fracVSA_SlogP11','fracVSA_SlogP12']

data_masterFinal_clean = data.select(final_column_order)

### Select data subset for use in training and testing

data_final = data.select(columns_labelsAndFeatures_subset)

data_final = data_masterFinal_clean

(training,testing) = data_final.randomSplit([0.5,0.5])

# Notebook Agenda, step 1:
* <font color='green'>__Step 1.__</font> verify this approach is feasible using RFC w/ NLP features (__"Model 1"__) followed by RFR w/ Numerical features (__"Model 2"__) <br>

* Note: labels are
  *	'label_q0' (Regression)
  *	'label_cat0' (5-class Classification)
  *	'label_cat1' (3-class Classification)
  *	'label_cat2' (4-class Classification)
  *	'label_cat3' (5-class Classification) 
  *	'label_cat4' (5-class Classification)


(training,testing) = data_final.randomSplit([0.5,0.5])

from pyspark.ml.classification import RandomForestClassifier
from pyspark.ml.evaluation import MulticlassClassificationEvaluator

featureChoices = ['FEAT_f_all_cv2', 'FEAT_f_all_cv2_idf', 'FEAT_f_all_cv5','FEAT_f_all_n2g_cv2', 'FEAT_f_all_n2g_cv5', 
                'FEAT_f_subset_cv2', 'FEAT_f_subset_cv2_idf', 'FEAT_f_subset_cv5', 'FEAT_f_subset_n2g_cv2', 'FEAT_f_subset_n2g_cv5', 
                'FEAT_f_subset2_cv2', 'FEAT_f_subset2_cv2_idf', 'FEAT_f_subset2_cv5', 'FEAT_f_subset2_n2g_cv2', 'FEAT_f_subset2_n2g_cv5']

featuresName = 'FEAT_f_subset_cv2'#'FEAT_all_n2g_cv2' #'FEAT_subset2_cv5'
labelName = 'label_cat1' #'label_cat4'

rfc = RandomForestClassifier(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction_1',numTrees=80)
mcEvaluator = MulticlassClassificationEvaluator(labelCol=labelName, predictionCol="prediction_1", metricName="accuracy")

mymodel_rfc = rfc.fit(training)
myresults_rfc = mymodel_rfc.transform(testing)

model_eval_rfc = mcEvaluator.evaluate(myresults_rfc)

print(f"RFC accuracy = {model_eval_rfc}")

from pyspark.ml.regression import RandomForestRegressor
from pyspark.ml.evaluation import RegressionEvaluator

featuresName = 'FEAT_rdkit_1b'#'FEAT_rdkit_2bANOVA'#
labelName = 'label_q0' 

rfr = RandomForestRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction_2')
regEvaluator = RegressionEvaluator(labelCol=labelName,predictionCol='prediction_2',metricName='r2')

mymodel_rfr1 = rfr.fit(training)
myresults_rfr1 = mymodel_rfr1.transform(myresults_rfc)

model_eval_rfr1 = regEvaluator.evaluate(myresults_rfr1)

print(f"RFR_1 r2 = {model_eval_rfr1}")

* let's try training an alternative regression model to RFR1, that incorporates the rfc prediction into its features

from pyspark.ml.feature import OneHotEncoder
from pyspark.ml.linalg import Vectors
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.feature import StandardScaler

#assembler = VectorAssembler( inputCols=['prediction_1','FEAT_rdkit_1b'],outputCol='features_RFR2')
#ohc = OneHotEncoder(inputCol='prediction_1',outputCol='prediction_1_enc') # new
#assembler = VectorAssembler( inputCols=['prediction_1_enc','FEAT_rdkit_1b'],outputCol='features_RFR2') # new

scaler = StandardScaler(inputCol='FEAT_rdkit_1b',outputCol='FEAT_rdkit_1bS')
ohc = OneHotEncoder(inputCol='prediction_1',outputCol='prediction_1_enc') # new
assembler = VectorAssembler( inputCols=['prediction_1_enc','FEAT_rdkit_1bS'],outputCol='features_RFR2')

#scaler = StandardScaler(inputCol='FEAT_rdkit_1b',outputCol='FEAT_rdkit_1bS')
#assembler = VectorAssembler( inputCols=['prediction_1','FEAT_rdkit_1bS'],outputCol='features_RFR2')

myresults_rfc = scaler.fit(myresults_rfc).transform(myresults_rfc)
myresults_rfc = ohc.fit(myresults_rfc).transform(myresults_rfc)
combined_model_data = assembler.transform(myresults_rfc)

(combined_training,combined_testing) = combined_model_data.randomSplit([0.7,0.3])

from pyspark.ml.regression import RandomForestRegressor,LinearRegression,GBTRegressor
from pyspark.ml.evaluation import RegressionEvaluator

featuresName = 'features_RFR2'
labelName = 'label_q0' 

rfr2 = RandomForestRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='combinedPrediction_1')#,numTrees=100)
#rfr2 = GBTRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='combinedPrediction_1')
regEvaluator2 = RegressionEvaluator(labelCol=labelName,predictionCol='combinedPrediction_1',metricName='r2')

mymodel_rfr2 = rfr2.fit(combined_training)
myresults_rfr2 = mymodel_rfr2.transform(combined_testing)

model_eval_rfr2 = regEvaluator2.evaluate(myresults_rfr2)

print(f"RFR_2 r2 = {model_eval_rfr2}")

* let's try just training a separate model to analyze the first two independent predictions (rfc and rfr1) only

from pyspark.ml.linalg import Vectors
from pyspark.ml.feature import VectorAssembler

assembler = VectorAssembler( inputCols=['prediction_1','prediction_2'],outputCol='features_RFR3')
combined_model_data = assembler.transform(myresults_rfr1)

(combined_training,combined_testing) = combined_model_data.randomSplit([0.5,0.5])

from pyspark.ml.regression import RandomForestRegressor
from pyspark.ml.evaluation import RegressionEvaluator

featuresName = 'features_RFR3'
labelName = 'label_q0' 

rfr3 = RandomForestRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction_3')
regEvaluator3 = RegressionEvaluator(labelCol=labelName,predictionCol='prediction_3',metricName='r2')

mymodel_rfr3 = rfr3.fit(combined_training)
myresults_rfr3 = mymodel_rfr3.transform(combined_testing)

model_eval_rfr3 = regEvaluator3.evaluate(myresults_rfr3)

print(f"RFR_3 r2 = {model_eval_rfr3}")

from pyspark.ml.regression import GBTRegressor
from pyspark.ml.evaluation import RegressionEvaluator

gbtr = GBTRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')


featuresName = 'features_RFR3'
labelName = 'label_q0' 

rfr3 = GBTRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction_3')
regEvaluator3 = RegressionEvaluator(labelCol=labelName,predictionCol='prediction_3',metricName='r2')

mymodel_rfr3 = rfr3.fit(combined_training)
myresults_rfr3 = mymodel_rfr3.transform(combined_testing)

model_eval_rfr3 = regEvaluator3.evaluate(myresults_rfr3)

print(f"RFR_3 r2 = {model_eval_rfr3}")

# <font color='red'> Prior Scratch Work </font>

''' # create a history log for evaluation trials
'''
evaluation_history = {}

import pickle
with open('evaluation_history_temp.pickle', 'wb') as handle:
    pickle.dump(evaluation_history, handle, protocol=pickle.HIGHEST_PROTOCOL)


# Note: to load eval history, use:
# with open('evaluation_history.pickle', 'rb') as handle:
#    evaluation_history = pickle.load(handle)

(training,testing) = data_final.randomSplit([0.7,0.3])

trial = 0
trial_description = 'regression performance test: best NLP & RDkit features combined (frags_all CV with RDKit features_1b)'
featuresName = 'features' 
labelName = 'label_q0'  

'''# Import Regression Models
'''
from pyspark.ml.regression import (LinearRegression,
                                   DecisionTreeRegressor,RandomForestRegressor,GBTRegressor,
                                   GeneralizedLinearRegression,IsotonicRegression)

lr = LinearRegression(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')
dtr = DecisionTreeRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')
rfr = RandomForestRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')
gbtr = GBTRegressor(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')
glr = GeneralizedLinearRegression(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')
ir = IsotonicRegression(featuresCol=featuresName,labelCol=labelName,predictionCol='prediction')

models_regression = {'lr':lr,'dtr':dtr,'rfr':rfr,'gbtr':gbtr,'glr':glr,'ir':ir}


''' # Create DataFrame for displaying eval metrics of current trial
'''
eval_df = pd.DataFrame() # already exists from prior run


''' # Specify experiment type
'''
modeltype = lr  # --- SPECIFY!!! (lr,dtr,rfr,gbtr,glr,ir)
modelname = 'lr'
modelname_short = 'lr' # SPECIFY
iteration = 0


''' # Load eval History file for logging eval metrics of all trials
'''
import pickle
with open('evaluation_history.pickle', 'rb') as handle:
    evaluation_history = pickle.load(handle)

trialName = f"trial_{trial}"
iterationName = f"iteration_{iteration}"
evaluation_history[trialName] = {}
evaluation_history[trialName]['features_set'] = featuresName
evaluation_history[trialName]['description'] = trial_description
evaluation_history[trialName][modelname_short] = {}
evaluation_history[trialName][modelname_short][iteration] = {}
evaluation_history[trialName][modelname_short][iteration]['label'] = labelName


''' # BEGIN TRIAL
'''


evaluation_history[trialName][modelname_short][iteration] = {}

mymodel = modeltype.fit(training)
myresults = mymodel.transform(testing)

# CALCULATE KEY EVALS
from pyspark.ml.evaluation import RegressionEvaluator
regEvaluator = RegressionEvaluator(labelCol=labelName,predictionCol='prediction')

evaluator = regEvaluator
evalMetrics = {regEvaluator:['rmse','mse','mae','r2','var']}

evaluation = []

for each_metric in evalMetrics[evaluator]:        
    metric = each_metric

    result = evaluator.evaluate(myresults, {evaluator.metricName: metric})

    evaluation.append((metric,result))

    evaluation_history[trialName][modelname_short][iteration][metric] = result

#r2_adj = mymodel.summary.r2adj
#evaluation.append(('r2_adj(Training)',r2_adj))
column0 = [x for x,y in evaluation]
column1 = [y for x,y in evaluation]
eval_df['metric'] = column0
eval_df[modelname] = column1

eval_df.head(10)

eval_df.head(10)

''' # BACKUP EVALUATION HISTORY
'''
import pickle
with open('evaluation_history.pickle', 'wb') as handle:
    pickle.dump(evaluation_history, handle, protocol=pickle.HIGHEST_PROTOCOL)

# iteration 5
pd.set_option('display.max_columns',None)
eval_df

# iteration 4
pd.set_option('display.max_columns',None)
eval_df

### Results:
**Findings:**
1. the most effective fragment representation is `frags_all`
2. the most effective nlp representation is **CountVectorizer** or CV-IDF > bigram CV > word2vec

## NEW: Train & Evaluate `frags_all` (CV, bigram CV, word2vec) using different regression models
**Goal:** 
1. Determine the most useful type of regression model for this data

# iteration 3
eval_df

* Observations:
  * at least when sourced from "frags_all", the most effective vector features for the fragments are:
    * cv-idf > tf-idf > 2-gram-cv-idf > word2vec
    * sometimes, word2vec has been better than bigram cv-idf
    * default settings word2vec usually perform best
  * it might be interesting to **check how well bigrams perform in a word2vec model**

''' # BACKUP EVALUATION HISTORY
# --- NOTE: to load eval history, use:
# with open('evaluation_history.pickle', 'rb') as handle:
#    evaluation_history = pickle.load(handle)
'''
import pickle

with open('evaluation_history.pickle', 'wb') as handle:
    pickle.dump(evaluation_history, handle, protocol=pickle.HIGHEST_PROTOCOL)


''' 

'''

eval_df

testx = eval_df.set_index('metric')
testx = testx.transpose()
testx.head(6)

''' # Check feature importances
'''
import pandas as pd
for k, v in test_subset_final.schema["features_tf_idf"].metadata["ml_attr"]["attrs"].items():
    features_df = pd.DataFrame(v)
    # format into a pandas dataframe and display
    rf_output = rfmodel.featureImportances
    features_df['Importance'] = features_df['idx'].apply(lambda x: rf_output[x] if x in rf_output.indices else 0)

features_df.sort_values("Importance", ascending=False, inplace=True)
features_df.head()


* NOTE: to get vocab back, you can do it from the CountVectorizer indices
  * fit CountVectorizer separately, store the vocab like this:
    * `vectorizer = CountVectorizer(inputCol="tokens", outputCol="features").fit(df)`
    * `vectorizer.vocabulary`

## Test: Check feature importances of fragments as features 

# <font color='red'> Proceed with other feature vectors </font>

''' 
# FEATURE SELECTION:
'''
# to load the Features Information, use the command:
featuresDF = pd.read_parquet('featuresCatalogDF_2022-08-16.parquet')

#index_pos = featuresDF[featuresDF['name']=='F1a'].index[0]
feature_set1a = featuresDF.loc[0,'features']
feature_set1b = featuresDF.loc[1,'features']
feature_set2a = featuresDF.loc[2,'features']
feature_set2b = featuresDF.loc[3,'features']
feature_set3  = featuresDF.loc[4,'features']
feature_set4a = featuresDF.loc[5,'features']
feature_set4b = featuresDF.loc[6,'features']


# VECTOR ASSEMBLY - feature sets 1a,1b,2a,2b,3,4a,4b
from pyspark.ml.linalg import Vector
from pyspark.ml.feature import (VectorAssembler,VectorIndexer)

vec_assembler1a = VectorAssembler(inputCols = feature_set1a, outputCol='features1a')
vec_assembler1b = VectorAssembler(inputCols = feature_set1b, outputCol='features1b')
vec_assembler2a = VectorAssembler(inputCols = feature_set2a, outputCol='features2a')
vec_assembler2b = VectorAssembler(inputCols = feature_set2b, outputCol='features2b')
vec_assembler3 = VectorAssembler(inputCols = feature_set3, outputCol='features3')
vec_assembler4a = VectorAssembler(inputCols = feature_set4a, outputCol='features4a')
vec_assembler4b = VectorAssembler(inputCols = feature_set4b, outputCol='features4b')

from pyspark.ml import Pipeline
feature_pipeline = Pipeline(stages=[vec_assembler1a,
                                    vec_assembler1b,
                                    vec_assembler2a,
                                    vec_assembler2b,
                                    vec_assembler3,
                                    vec_assembler4a,
                                    vec_assembler4b])
data_features = feature_pipeline.fit(data).transform(data)


''' 
# LABELS 
# Data has 5 labels; 1 continuous BA percentage, 4 discretized groups.
#  - ranges of discretized groups were selected by examining histogram distributions, mean & stdev of BA %
#  - we'll add one more discretized label alternative using Spark's QuantileDiscretizer, into 5 groups

# -- Add QuantileDiscretizer labels
from pyspark.ml.feature import QuantileDiscretizer
import pandas as pd
qd5 = QuantileDiscretizer(numBuckets=5,inputCol='BA_pct',outputCol='label_QD5')

data_features = qd5.fit(data_features).transform(data_features)

# -- INDEX / ENCODE LABELS
from pyspark.ml.feature import (StringIndexer,OneHotEncoder)

label_quant0 = 'BA_pct'
label_cat0_vector = OneHotEncoder(inputCol='label_QD5',outputCol='label_cat0_vector')

label_cat1_index = StringIndexer(inputCol='label1',outputCol='label_cat1_index')
label_cat1_vector = OneHotEncoder(inputCol='label_cat1_index',outputCol='label_cat1_vector')

label_cat2_index = StringIndexer(inputCol='label2',outputCol='label_cat2_index')
label_cat2_vector = OneHotEncoder(inputCol='label_cat2_index',outputCol='label_cat2_vector')

label_cat3_index = StringIndexer(inputCol='label3a',outputCol='label_cat3_index')
label_cat3_vector = OneHotEncoder(inputCol='label_cat3_index',outputCol='label_cat3_vector')

label_cat4_index = StringIndexer(inputCol='label3b',outputCol='label_cat4_index')
label_cat4_vector = OneHotEncoder(inputCol='label_cat4_index',outputCol='label_cat4_vector')

from pyspark.ml import Pipeline
label_pipeline = Pipeline(stages=[label_cat0_vector,
                                 label_cat1_index,label_cat1_vector,
                                 label_cat2_index,label_cat2_vector,
                                 label_cat3_index,label_cat3_vector,
                                 label_cat4_index,label_cat4_vector])
'''
data_features = data_features.select(['Name','BA_pct','label_QD5','label1','label2','label3a','label3b','features1a','features1b','features2a','features2b','features3','features4a','features4b'])

#data_prefinal = label_pipeline.fit(data_features).transform(data_features)

data_prefinal2 = data_prefinal.withColumnRenamed('BA_pct','label_q0')
data_prefinal2 = data_prefinal2.withColumnRenamed('label_QD5','label_cat0')
data_prefinal2 = data_prefinal2.withColumnRenamed('label_cat1_index','label_cat1')
data_prefinal2 = data_prefinal2.withColumnRenamed('label_cat2_index','label_cat2')
data_prefinal2 = data_prefinal2.withColumnRenamed('label3a','label3')
data_prefinal2 = data_prefinal2.withColumnRenamed('label3b','label4')
data_prefinal2 = data_prefinal2.withColumnRenamed('label_cat3_index','label_cat3')
data_prefinal2 = data_prefinal2.withColumnRenamed('label_cat4_index','label_cat4')

data_final = data_prefinal2.select(['Name','label_q0',
                                    'label_cat0','label_cat1','label_cat2','label_cat3','label_cat4',
                                    'features1a','features1b','features2a','features2b','features3','features4a','features4b'])
